pyproject.toml
[project]
name = "cortana"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "coloredlogs>=15.0.1",
    "httpx>=0.28.1",
    "py-cord>=2.6.1",
    "pyyaml>=6.0.3",
]

[dependency-groups]
dev = [
    "python-dotenv>=1.2.1",
]
